backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
pydantic-settings==2.10.1
python-dotenv==1.1.1
SQLAlchemy==2.0.43
pymysql==1.1.1
cryptography==44.0.2
bcrypt==4.3.0
PyJWT==2.10.1
paho-mqtt==2.1.0

